rename service cache to registry cache
